my-sdk/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "my_sdk/__init__.py"

[tool.hatch.build.targets.wheel]
package = [
    "my_sdk",
]

[project]
name = "my-sdk"
description = "A simple SDK for demonstration purposes."
dynamic = ["version"]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "apache-airflow>=3.0.0",
    "typing_extensions>=4.0.0",
    ]

[project.entry-points."apache_airflow_provider"]
provider_info = "my_sdk:get_provider_info"